feat(appointment): 약속 결과 페이지 컴포넌트 분리
